SERVER/requirements.txt
flask
flask_cors
transformers
torch
Pillow
requests